Mensajes chat a la izquierda y derecha

+++ app/src/main/java/com/ignacio_natalia/puzzleteca/modelos/chat/MensajeChat.java
@@ -4,15 +4,16 @@ public class MensajeChat {
 
     private Long idMensaje;
     private Long idConversacion;
-    private Long idUsuario;
+    private Integer idUsuario;
     private String nombre;
     private String contenido;
     private String creadoEn;
 
     public Long getIdMensaje() { return idMensaje; }
     public Long getIdConversacion() { return idConversacion; }
-    public Long getIdUsuario() { return idUsuario; }
+    public Integer getIdUsuario() { return idUsuario; }
     public String getNombre() { return nombre; }
     public String getContenido() { return contenido; }
     public String getCreadoEn() { return creadoEn; }
+
 }
